+++ rdfframework/rdfframework/forms/widgets/__init__.py
@@ -1,0 +1,3 @@
+__author__ = "Mike Stabile, Jeremy Nelson"
+
+from rdfframework.forms.widgets.formwidgets import *
