+++ rdfframework/rdfframework/processors/__init__.py
@@ -1,0 +1,4 @@
+from .rdfprocessors import *
+
+__author__ = "Mike Stabile, Jeremy Nelson"
+
