+++ rdfframework/rdfframework/sparql/jsonObjectQueryTemplate.rq
