+++ rdfframework/rdfframework/sparql/kds-vocab-summary.rq
