+++ rdfframework/rdfframework/sparql/sparqlItemTemplate.rq
