+++ rdfframework/rdfframework/turtle/kds-app.ttl
