+++ rdfframework/rdfframework/turtle/kds-vocab.ttl
