--- rdfframework/sparql/jsonApplicationDefaults.rq
@@ -1,77 +1,0 @@
-prefix acl: <http://www.w3.org/ns/auth/acl#> 
-prefix foaf: <http://xmlns.com/foaf/0.1/> 
-prefix kds: <http://knowledgelinks.io/ns/data-structures/> 
-prefix kdr: <http://knowledgelinks.io/ns/data-resources/> 
-prefix obi: <https://w3id.org/openbadges#> 
-prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> 
-prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> 
-prefix schema: <https://schema.org/> 
-prefix xsd: <http://www.w3.org/2001/XMLSchema#>
-
-SELECT (CONCAT("{",GROUP_CONCAT(?sections; SEPARATOR=","),"}") AS ?app)
-{
-GRAPH <http://knowledgelinks.io/ns/application-framework/>
-{
-  SELECT ?sections
-  {
-    {
-  SELECT ?klapp (count(?p)>1 AS ?obj) (GROUP_CONCAT(?oj; SEPARATOR=',') AS ?section)
-  {
-    {
-  SELECT ?klapp ?p  (GROUP_CONCAT(?ojf; SEPARATOR=',') AS ?oja) 
-                    {
-									SELECT ?klapp ?p ?o (CONCAT(IF(count(?o)>1,'{','"'),GROUP_CONCAT(?oj; SEPARATOR=','),IF(count(?o)>1,'}','"')) AS ?ojf)
-                                    {
-                                      {
-
-                                        SELECT  ?klapp ?p ?o ?bn_ ?p1_  (GROUP_CONCAT(?o1j; SEPARATOR=',') AS ?o1jf)
-                                                      {
-                                                        {
-                                                          ?klapp rdfs:subClassOf kds:InternalClass .
-                                                          ?bns a ?klapp .
-                                                        } UNION {
-                                                          ?klapp rdfs:subClassOf kds:FormClass .
-                                                          ?bns a ?klapp .
-                                                        } UNION {
-	                                                        ?klapp a kds:Application .
-	                                                        ?klapp ?ps ?os .
-                                                        }
-                                                        BIND(IF(bound(?bns),?bns,?os) AS ?o).
-                                                        BIND(IF(bound(?bns),?bns,?ps) AS ?p).
-                                                        OPTIONAL {
-                                                          {?o ?p1 ?o1 .
-                                                          }
-                                                        } .
-                                                        OPTIONAL {
-                                                          {?o ?p1 ?o1 .
-                                                          FILTER(ISBLANK(?o))}
-                                                        } .
-                                                        
-                                                        BIND(IF(bound(?o1),CONCAT('"',STR(?o1),'"'),"") AS ?o1j)
-                                                        BIND(IF(BOUND(?p1),?p1,"") AS ?p1_) .
-                                                        BIND(IF(BOUND(?o),?o,"") AS ?bn_).
-                                                        
-                                                      }
-                                                     GROUP BY ?klapp ?p ?o ?bn_ ?p1_ 
-													}
-                                      BIND(IF(CONTAINS(?o1jf,'","'),true,false) AS ?bracket)
-                                      BIND(IF(isblank(?o)||?o=?p,CONCAT('"',REPLACE(STR(?p1_),"^(.*[#/])", ""),'":',IF(?bracket,'[',""),?o1jf,IF(?bracket,']',"")),?o) AS ?oj)
-                                    }
-                                    GROUP BY ?klapp ?p ?o 
-								  			
-           }
-                    GROUP BY ?klapp ?p 
-
- }
-    BIND(IF(CONTAINS(?oja,'},{'),true,false) AS ?bracket)
-    #BIND(CONCAT(IF(?bracket,'[',""),?oja,IF(?bracket,']',"")) AS ?oj)
-    BIND(CONCAT('"',REPLACE(STR(?p),"^(.*[#/])", ""),'":',IF(?bracket,'[',""),?oja,IF(?bracket,']',"")) AS ?oj)
-  }
-GROUP BY ?klapp
-}
-      BIND(CONCAT('"',REPLACE(STR(?klapp),"^(.*[#/])", ""),'":',IF(?obj,'{',""),?section,IF(?obj,'}',"")) AS ?sections)
-}
-								  			
-}
-}
-
